requirements.txt
beautifulsoup4
lxml
gutenbergpy
EbookLib
requests
python-dotenv
openai